Crickbot
It is an Image Processing based project and the processing was done in
MATLAB. This project was basically about a robot playing cricket. The
overhead camera above the cricket field captured video continuously, and
the processing was done on the input video stream. An algorithm was made
for the robot to move to the ball which was thrown into the field,
capture it and bring it to the deposit point. The robot had different
colored geometrical shapes on it so as to differentiate it from the
field and these shapes also help in processing and moving robot to the
different position. The robot worked autonomously depending on the
command given by MATLAB after processing the live video input. This was
my first image processing based project. This project included coding
skills, mechanical skills for designing the robot, basic mathematics
concepts for designing algorithm for how robot will move to the ball and
back to deposit point.